+++ webman-amplifier.php
@@ -3,7 +3,7 @@
  * Plugin Name:  WebMan Amplifier
  * Plugin URI:   https://github.com/webmandesign/webman-amplifier/
  * Description:  Pack of additional WordPress features. Contains additional custom post types, shortcodes, page builder integration, meta box generator and icon font management.
- * Version:      1.6.0
+ * Version:      1.6.1
  * Author:       WebMan Design, Oliver Juhas
  * Author URI:   https://www.webmandesign.eu/
  * License:      GPL-3.0-or-later
@@ -32,7 +32,7 @@ defined( 'ABSPATH' ) || exit;
 	// Define global constants
 
 		// Plugin version
-		if ( ! defined( 'WMAMP_VERSION' ) ) define( 'WMAMP_VERSION', '1.6.0' );
+		if ( ! defined( 'WMAMP_VERSION' ) ) define( 'WMAMP_VERSION', '1.6.1' );
 
 		// Paths
 		if ( ! defined( 'WMAMP_PLUGIN_FILE' ) )  define( 'WMAMP_PLUGIN_FILE',  __FILE__                                          );